Move Reservoir_Outlet to dedicated file.
Moving Reservoir_Outlet from within Nonlinear_Reservoir.hpp to new,
dedicated header file Reservoir_Outlet.hpp for more compartmentalized
design.
